temporarrily fixed the monotonicity violation (pdf can be >1) by using pdf(x)/pdf(mean) as probability. fixed another critical bug  in PlaneSeg

+++ cfg3d/src/helper.cpp
@@ -47,7 +47,7 @@ void initializeTerminals(int & maxSegNum) {
     }
     
     Terminal::totalNumTerminals=maxSegNum;
-    assert(terminals.size()==Terminal::totalNumTerminals);
+    assert((int)terminals.size()==Terminal::totalNumTerminals);
     
 }
